Revert "refactor validators"
This reverts commit 8929198af593ad197e7fdf070a8932eb8f8b13d4.
